requirements.txt
langchain
google-cloud-aiplatform
streamlit
langchain-google-vertexai